added students Repository

+++ src/main/java/bbw/ch/Morning/Repository/StudentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface StudentRepository extends CrudRepository<StudentEntity, Integer> {
+public interface StudentRepository extends CrudRepository<StudentEntity, Long> {
     Optional<StudentEntity> findByEmail(String findByEmail);
 }
